+++ Astra/astra/pages/about_page.h
@@ -1,0 +1,7 @@
+#ifndef ASTRA_ABOUT_PAGE_H
+#define ASTRA_ABOUT_PAGE_H
+
+namespace astra { class Menu; }
+void AboutPage_AddToMenu(astra::Menu *menu);
+
+#endif