+++ Astra/astra/pages/display_page.h
@@ -1,0 +1,5 @@
+#ifndef ASTRA_DISPLAY_PAGE_H
+#define ASTRA_DISPLAY_PAGE_H
+namespace astra { class Menu; }
+void DisplayPage_AddToMenu(astra::Menu *menu);
+#endif